ADA-73 Refactor getting the semesters for the current academic year
